[debugger] fixed save all command problem

+++ tools/debug/Editor/EditorManager.py
@@ -124,19 +124,20 @@ class EditorManager(QWidget):
 		if result > 0:
 			self.ui.directory.setText(path)
 
-    def save(self):
+    def save(self, editor_index = None):
 		
 		editor = None
-		editor = self.app.focusWidget()
+		if editor_index is not None:
+			editor = self.tab.editors[editor_index]
+		else:
+			editor = self.app.focusWidget()
 
 		if isinstance(editor, Editor):
-			if editor.tempfile == False:
-				currentText = open(editor.path).read()
-			else:
-				self.saveas()
+			if editor.tempfile is True:
+				self.saveas(editor_index)
 				return 
 
-			editor.save()
+			editor.save(None, editor_index)
 
 			if self.editors.has_key(editor.path):
 			    self.editors[editor.path][2] = editor.text() 
@@ -146,12 +147,20 @@ class EditorManager(QWidget):
     def saveall(self):
 		index = 0 
 		for editor in self.tab.editors :
+			if self.tab.editors[index].tempfile is True :
+			    index_offset = 0
+			else :
+			    index_offset = 1
 			self.save(index)
-			index = index + 1
+			index = index + index_offset
 
-    def close(self):
-		index = self.tab.currentIndex()
-		self.tab.closeTab(index)
+    def close(self, editor_index = None):
+        if editor_index is None:
+            index = self.tab.currentIndex()
+        else:
+            index = editor_index
+
+        self.tab.closeTab(index)
 
     def get_bp_path(self, file_path):
         if re.search(str(self.deviceManager.path()), file_path) is not None :
@@ -183,12 +192,15 @@ class EditorManager(QWidget):
                 else:
                     idx += 1
 
-    def saveas(self):
+    def saveas(self, editor_index= None):
 		self.dialog = QDialog()
 		self.ui = Ui_saveAsDialog()
 		self.ui.setupUi(self.dialog)
 
-		editor = self.app.focusWidget()
+		if editor_index is not None:
+			editor = self.tab.editors[editor_index]
+		else:
+			editor = self.app.focusWidget()
 
 		cur_file = re.search('(\w+)[.](\w+)', editor.path).group()
 		cur_dir = editor.path[:re.search('(\w+)[.](\w+)', editor.path).start()-1]
@@ -203,12 +215,23 @@ class EditorManager(QWidget):
 			cur_file = self.ui.filename.text() 
 			new_path = cur_dir+'/'+cur_file
 
+			if str(new_path) == str(editor.path) and editor.tempfile == False :
+			    msg = QMessageBox()
+			    msg.setText('The file named "' + cur_file + '" already exists in "'+ cur_dir + '". Do you want to replace it? ')
+			    msg.setInformativeText('Saving it will overwrite its contents.')
+			    msg.setStandardButtons(QMessageBox.Save | QMessageBox.Cancel)
+			    msg.setDefaultButton(QMessageBox.Cancel)
+			    msg.setWindowTitle("Warning")
+			    ret = msg.exec_()
+			    if ret == QMessageBox.Cancel:
+			        return 
+
 			print "[VDBG] Save As .. ' %s '"%new_path
 			bp_file = self.get_bp_path(new_path)
 
 			if editor.tempfile == False :
 				currentText = editor.text()
-				editor.save(self.editors[editor.path][2])
+				editor.save(self.editors[editor.path][2], editor_index)
 				editor.set_temp_marker(bp_file)
 				editor.delete_marker()
 				self.close()
@@ -220,7 +243,7 @@ class EditorManager(QWidget):
 				"""
 				self.newEditor(new_path, None, None, None, False, None, False, currentText)
 				editor = self.app.focusWidget()
-				editor.save()
+				editor.save(None, editor_index)
 				editor.add_marker()
 			else :
 				currentText = editor.text()
@@ -230,7 +253,9 @@ class EditorManager(QWidget):
 				editor.set_temp_marker(bp_file)
 				if editor.path != new_path :
 				    editor.delete_marker()
-				self.close()
+				
+				self.close(editor_index)
+
 				"""
 				if len(self.bp_info[1]) > 0:
 				    for i in range (0, len(self.bp_info[1])) :
@@ -260,7 +285,8 @@ class EditorManager(QWidget):
         editor.tempfile = tempfile
         editor.fileIndex = fileIndex
 
-        closedTab = None
+        #closedTab = None
+        closedTab = []
 
         # Default to opening in the first tab group
         tabGroup = 0
@@ -272,13 +298,19 @@ class EditorManager(QWidget):
             self.tab = self.EditorTabWidget(self.splitter)
             self.editorGroups.append(self.tab) 
         # If the file is already open, just use the open document
+
         if self.editors.has_key(path):
             for k in self.editors:
-				if not k in self.tab.paths:
-					closedTab = k
+				if not str(k) in self.tab.paths:
+					#closedTab = k
+					closedTab.append(k)
 
-            if closedTab != None:
-        		self.editors.pop(closedTab)
+            #if closedTab != None:
+            if len(closedTab) > 0 :
+        		#self.editors.pop(closedTab)
+        		for c_idx in range (0, len(closedTab)):
+        		    self.editors.pop(closedTab[c_idx])
+
         		for k in self.tab.paths:
 					self.editors[k][1] = self.tab.paths.index(k) 
         		if editor.tempfile == False:
@@ -287,8 +319,8 @@ class EditorManager(QWidget):
         		    else:
         			    editor.setText(currentText) 
 
-
-            if closedTab != path:
+            #if closedTab != path:
+            if not path in closedTab :
 				for k in self.editors:
 					if path == k:
 						curIndex = self.editors[k][1]
@@ -312,7 +344,8 @@ class EditorManager(QWidget):
 										self.currentEditor.markerAdd(self.currentEditor.current_line, 
 																	Editor.DEACTIVE_BREAK_MARKER_NUM)
 								elif prev_file != None:
-									if closedTab != prev_file:
+									#if closedTab != prev_file:
+									if not prev_file in closedTab :
 										for j in self.editors:
 											if prev_file == j:
 												prevIndex = self.editors[j][1]
@@ -338,6 +371,7 @@ class EditorManager(QWidget):
 										self.currentEditor.markerDelete(nextCurLine, Editor.DEACTIVE_BREAK_MARKER_NUM)
 										self.currentEditor.markerAdd(nextCurLine, Editor.ARROW_DEACTIVE_BREAK_MARKER_NUM)
 									self.currentEditor.current_line = nextCurLine
+				#print ("return here - opened tab")
 				return self.currentEditor
         elif tempfile == False:
             path = os.path.join (self.deviceManager.path(), path)
@@ -430,6 +464,7 @@ class EditorManager(QWidget):
                 else :
                     self.main.ui.actionUndo.setEnabled(False)
 
+        #print ("return there - new tab")
         return editor
 
     @pyqtSlot()
